requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
pillow>=10.1.0
opencv-python>=4.9.0
schedule>=1.2.0
pydantic>=2.5.0
gitpython>=3.1.40
